Move views to the "view" package and modify them to use a common "BasePane" class
Moved:
    EditorPane -> view.EditorPane
    ReviewerPane -> view.ReviewerPane
    LoginPane -> view.LoginPane
    ResearcherPane -> view.ResearcherPane

Updated those 4 panes to use BasePane as their base class and updated them to only set the center of the BorderPane (which BasePane) is based on, leaving the left to always be for navigation and the top to always be for the current page title.

Created NavigationPane which is embeded in the left of all pages via the BasePane, along with a convention to add new pages/panes to the nav list.